+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Display `<redtext` in posts.
 - Added CSS variables for greentext and redtext colors.
 
+### Fixed
+
+- Fixed #️⃣ emoji reaction.
+
 ## [5.1.0] - 2026-04-08
 
 ### Added